+++ src/libopentld/tld/MedianFlowTracker.cpp
@@ -70,11 +70,11 @@ void MedianFlowTracker::track(const Mat &prevMat, const Mat &currMat, Rect *prev
         int success = fbtrack(&prevImg, &currImg, bb_tracker, bb_tracker, &scale);
 
         //Extract subimage
-        float x, y, w, h;
-        x = floor(bb_tracker[0] + 0.5);
-        y = floor(bb_tracker[1] + 0.5);
-        w = floor(bb_tracker[2] - bb_tracker[0] + 1 + 0.5);
-        h = floor(bb_tracker[3] - bb_tracker[1] + 1 + 0.5);
+        int x, y, w, h;
+        x = static_cast<int>(floor(bb_tracker[0] + 0.5));
+        y = static_cast<int>(floor(bb_tracker[1] + 0.5));
+        w = static_cast<int>(floor(bb_tracker[2] - bb_tracker[0] + 1 + 0.5));
+        h = static_cast<int>(floor(bb_tracker[3] - bb_tracker[1] + 1 + 0.5));
 
         //TODO: Introduce a check for a minimum size
         if(!success || x < 0 || y < 0 || w <= 0 || h <= 0 || x + w > currMat.cols || y + h > currMat.rows || x != x || y != y || w != w || h != h) //x!=x is check for nan
